Move orchestrator and checkpoints to ddev/ai/runtime and move resources to their own files
