+++ ddev/tests/ai/phases/test_base.py
@@ -7,9 +7,9 @@ from datetime import UTC, datetime
 import pytest
 
 from ddev.ai.phases.base import FlowContext, Phase, PhaseOutcome
-from ddev.ai.phases.checkpoint import CheckpointManager
 from ddev.ai.phases.config import PhaseConfig
 from ddev.ai.phases.messages import PhaseFailedMessage, PhaseTrigger
+from ddev.ai.runtime.checkpoints import CheckpointManager
 from ddev.event_bus.exceptions import HookName, MessageProcessingError, ProcessorHookError
